Add files via upload
developed in Android Studio, using XML for the user interface and Java
for the backend logic. The Firebase library was integrated as the primary backend technology,
offering services such as real-time database management and user authentication. Firebase’s robust
and scalable infrastructure allowed for efficient data handling and user management, crucial for
managing subscriptions and interactions between users and TSPs.